requirements.txt
alabaster==0.7.12
asgiref==3.2.10
astroid==2.4.2
Babel==2.8.0
cachetools==4.1.0
certifi==2020.6.20
chardet==3.0.4
commonmark==0.9.1
Django==3.0.7
djangorestframework==3.11.0
docutils==0.16
google-api-core==1.21.0
google-auth==1.18.0
google-cloud-core==1.3.0
google-cloud-storage==1.29.0
google-cloud-texttospeech==2.1.0
google-cloud-translate==2.0.1
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpcio==1.29.0
idna==2.9
imagesize==1.2.0
isort==4.3.21
Jinja2==2.11.2
lazy-object-proxy==1.4.3
libcst==0.3.6
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==20.4
proto-plus==1.0.0
protobuf==3.12.2
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.7
pytz==2020.1
PyYAML==5.3.1
recommonmark==0.6.0
requests==2.24.0
rinoh-typeface-dejavuserif==0.1.1
rinoh-typeface-texgyrecursor==0.1.1
rinoh-typeface-texgyreheros==0.1.1
rinoh-typeface-texgyrepagella==0.1.1
rinohtype==0.4.0
rsa==4.6
six==1.15.0
snowballstemmer==2.0.0
Sphinx==1.7.5
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.2.2
sqlparse==0.3.1
toml==0.10.1
typing-extensions==3.7.4.2
typing-inspect==0.6.0
urllib3==1.25.9
wrapt==1.12.1
